+++ tests/verify-modules.js
@@ -97,6 +97,7 @@
     '/js/onboarding-view-runtime.js',
     '/js/onboarding-view.js',
     '/js/data-merge-lab-entries.js',
+    '/js/data-view-controls.js',
     '/js/pii-review.js',
     '/js/marker-detail-modal.js',
     '/js/marker-detail-modal-impl.js',
