+++ tsconfig.json
@@ -125,6 +125,7 @@
     "js/data-merge-lab-entries.js",
     "js/data-merge.js",
     "js/data.js",
+    "js/data-view-controls.js",
     "js/dna.js",
     "js/dna-ui.js",
     "js/dna-runtime.js",
